lesson-18/dispute-chatbot/spikes/SPIKE-001_vrol_schema/requirements.txt
openapi-schema-validator
pyyaml
faker

